requirements.txt
jupyterlab
pandas
snowflake-cli-labs
snowflake-snowpark-python
streamlit
